Moved images to public
