+++ app/public/assets/buildings.json
