+++ app/public/assets/quadrants.json
