+++ app/public/assets/spaces.json
